update layout code
